backend_django/requirements.txt
Django>=4.2
djangorestframework
django-cors-headers

# add other backend deps as needed
requests
beautifulsoup4
